Repaired format specifier problem in GPS logging.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@2162 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/TAIGA/TAIGA/AppDelegate.m
@@ -78,7 +78,7 @@
     NSString* s = [formatter formatDegreesLatitude:location.coordinate.latitude
                                          longitude:location.coordinate.longitude
                                     metersAltitude:location.altitude];
-    DDLogInfo(@"GPS: %@ (%d m accuracy)", s, location.horizontalAccuracy);
+    DDLogInfo(@"GPS: %@ (%d m accuracy)", s, (int) location.horizontalAccuracy);
 }
 
 - (void) gdbMessageReceived:(NSNotification*)notification
